+++ src/Factory.php
@@ -7,8 +7,6 @@ namespace Phlib\Beanstalk;
 use Phlib\Beanstalk\Connection\Socket;
 use Phlib\Beanstalk\Exception\InvalidArgumentException;
 use Phlib\Beanstalk\Pool\Collection;
-use Phlib\Beanstalk\Pool\RoundRobinStrategy;
-use Phlib\Beanstalk\Pool\SelectionStrategyInterface;
 
 /**
  * @package Phlib\Beanstalk
@@ -23,23 +21,11 @@ class Factory
     public function createFromArray(array $config): ConnectionInterface
     {
         if (array_key_exists('host', $config)) {
-            $config = [
-                'server' => $config,
-            ];
-        }
-
-        if (array_key_exists('server', $config)) {
-            $server = $this->normalizeArgs($config['server']);
-            $connection = $this->create($server['host'], $server['port'], $server['options']);
-        } elseif (array_key_exists('servers', $config)) {
-            if (!isset($config['strategyClass'])) {
-                $config['strategyClass'] = RoundRobinStrategy::class;
-            }
-            $servers = $this->createConnections($config['servers']);
-            $strategy = $this->createStrategy($config['strategyClass']);
-            $connection = new Pool(new Collection($servers, $strategy));
+            $config = $this->normalizeArgs($config);
+            $connection = $this->create($config['host'], $config['port'], $config['options']);
         } else {
-            throw new InvalidArgumentException('Missing required server(s) configuration');
+            $connections = $this->createConnections($config);
+            $connection = new Pool(new Collection($connections));
         }
 
         return $connection;
@@ -50,24 +36,20 @@ class Factory
      */
     public function createConnections(array $servers): array
     {
+        if (empty($servers)) {
+            throw new InvalidArgumentException('Missing server configurations');
+        }
+
         $connections = [];
-        foreach ($servers as $server) {
-            if (array_key_exists('enabled', $server) && $server['enabled'] === false) {
+        foreach ($servers as $config) {
+            if (array_key_exists('enabled', $config) && $config['enabled'] === false) {
                 continue;
             }
-            $server = $this->normalizeArgs($server);
-            $connection = $this->create($server['host'], $server['port'], $server['options']);
+            $config = $this->normalizeArgs($config);
+            $connection = $this->create($config['host'], $config['port'], $config['options']);
             $connections[] = $connection;
         }
         return $connections;
-    }
-
-    public function createStrategy(string $class): SelectionStrategyInterface
-    {
-        if (!class_exists($class)) {
-            throw new InvalidArgumentException("Specified Pool strategy class '{$class}' does not exist.");
-        }
-        return new $class();
     }
 
     private function normalizeArgs(array $serverArgs): array